requirements.txt
SQLAlchemy==1.4.26
python-telegram-bot==13.8.1
python-dotenv==0.19.2
psycopg2-binary==2.9.1
